# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SkyHigh Core – Digital Check-In System API
  description: |
    REST API for the SkyHigh Airlines digital self-service check-in system.
    Handles seat selection, time-bound holds, baggage validation, payment gating,
    waitlist management, and abuse detection under high-concurrency conditions.

    ## Authentication
    All endpoints (except `/health` and `/metrics`) require a valid JWT in the
    `Authorization: Bearer <token>` header. The JWT is validated but not issued
    by this service (external IdP).

    **JWT Claims:**
    | Claim | Type   | Description        |
    |-------|--------|--------------------|
    | sub   | string | Passenger UUID     |
    | iat   | number | Issued-at (epoch)  |
    | exp   | number | Expiration (epoch) |

    ## Error Format
    All error responses follow the **RFC 7807 Problem Details** specification
    with `Content-Type: application/problem+json`.
  version: 1.0.0
  contact:
    name: SkyHigh Core Team
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/check-ins:
    post:
      operationId: startCheckIn
      tags: [Check-In]
      summary: Start check-in (select seat and hold)
      description: |
        Initiates the check-in process by selecting a seat and placing a time-bound hold.
        The hold lasts **120 seconds**. During this window, the passenger must confirm
        check-in (optionally adding baggage). If the hold expires, the seat is automatically
        released back to the availability pool or assigned to the next waitlisted passenger.

        **Concurrency:** Uses Redis distributed locks (Redlock) and CAS (compare-and-swap)
        pattern to guarantee exactly one passenger holds a seat at any time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartCheckInRequest"
            example:
              flightId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              seatId: "11111111-1111-1111-1111-111111111111"
      responses:
        "201":
          description: Check-in started, seat held for 120 seconds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResponse"
              example:
                id: "cccccccc-cccc-cccc-cccc-cccccccccccc"
                passengerId: "pppppppp-pppp-pppp-pppp-pppppppppppp"
                flightId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                seatId: "11111111-1111-1111-1111-111111111111"
                status: "IN_PROGRESS"
                holdExpiresAt: "2026-02-14T15:20:00Z"
                createdAt: "2026-02-14T15:18:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Flight or seat not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                flightNotFound:
                  summary: Flight does not exist
                  value:
                    type: "https://skyhigh.com/problems/flight-not-found"
                    title: "Flight Not Found"
                    status: 404
                    detail: "Flight with id 'a1b2c3d4-e5f6-7890-abcd-ef1234567890' was not found"
                    instance: "/api/v1/check-ins"
                seatNotFound:
                  summary: Seat does not exist on the flight
                  value:
                    type: "https://skyhigh.com/problems/seat-not-found"
                    title: "Seat Not Found"
                    status: 404
                    detail: "Seat with id '11111111-1111-1111-1111-111111111111' was not found on flight 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'"
                    instance: "/api/v1/check-ins"
        "409":
          description: Seat already held or passenger already checked in
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                seatAlreadyHeld:
                  summary: Seat is held by another passenger
                  value:
                    type: "https://skyhigh.com/problems/seat-already-held"
                    title: "Seat Already Held"
                    status: 409
                    detail: "Seat with id '11111111-1111-1111-1111-111111111111' is not available"
                    instance: "/api/v1/check-ins"
                alreadyCheckedIn:
                  summary: Passenger already has an active check-in
                  value:
                    type: "https://skyhigh.com/problems/already-checked-in"
                    title: "Already Checked In"
                    status: 409
                    detail: "Passenger already has an active check-in for this flight."
                    instance: "/api/v1/check-ins"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by an external IdP. Claims: `sub` (passenger UUID), `iat`, `exp`.

  # ── Parameters ────────────────────────────────
...
  schemas:
...
    CheckInStatus:
      type: string
      enum: [IN_PROGRESS, AWAITING_PAYMENT, COMPLETED, CANCELLED]
      description: Status of a check-in process
...
    StartCheckInRequest:
      type: object
      required: [flightId, seatId]
      properties:
        flightId:
          type: string
          format: uuid
          description: UUID of the flight to check in for
        seatId:
          type: string
          format: uuid
          description: UUID of the seat to hold
...
    CheckInResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        passengerId:
          type: string
          format: uuid
        flightId:
          type: string
          format: uuid
        seatId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/CheckInStatus"
        baggageWeight:
          type: number
          format: double
          nullable: true
          description: Baggage weight in kg
        excessFee:
          type: number
          format: double
          nullable: true
          description: Calculated excess baggage fee
        paymentId:
          type: string
          nullable: true
          description: External payment reference ID
        holdExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: When the seat hold expires (hold start + 120s)
        confirmedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when check-in was confirmed
        message:
          type: string
          nullable: true
          description: Human-readable status message (e.g., payment instructions)
        createdAt:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details error response
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://skyhigh.com/problems/seat-already-held"
        title:
          type: string
          description: Short human-readable summary of the problem type
          example: "Seat Already Held"
        status:
          type: integer
          description: HTTP status code
          example: 409
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "Seat 12A on flight SH-1042 is currently held by another passenger."
        instance:
          type: string
          description: URI reference of the specific occurrence (request path)
          example: "/api/v1/check-ins"

    # ── Health Schemas ────────────────────────
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://skyhigh.com/problems/bad-request"
            title: "Bad Request"
            status: 400
            detail: "page must not be less than 1"
            instance: "/api/v1/flights"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://skyhigh.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Missing or invalid authentication token."
            instance: "/api/v1/flights"
...
security:
  - BearerAuth: []